Moved existing pipes to new pipes folder
