+++ Singer.API/ClientApp/package.json
@@ -1,6 +1,6 @@
 {
    "name": "Singer",
-   "version": "0.1.0-issue-theme-update-dashboard.1213",
+   "version": "0.1.0-issue-theme-update-dashboard.1214",
    "scripts": {
       "ng": "ng",
       "start": "ng serve",
