+++ Singer.API/ClientApp/src/app/modules/core/Pipes/agegroup-to-color-pipe.pipe.spec.ts
