+++ Singer.API/ClientApp/src/app/modules/core/Pipes/keys.pipe.ts
